Allow ignoring refs when checking for global-refdb
Global refdb may not store one or more refs which makes sense
only on the local refdb.

Allows specifying a generic set of refs instead of hardcoding
a single one. This flexibility allows the consumer plugins
(e.g. multi-site or high-availability) to bind the set of refs
that should be ignored.

Also tidy up static imports of Collections constants and
methods and remove redundant static and public modifiers in
interfaces.

Set the new version to v3.3.1.

Bug: Issue 13971
Change-Id: Iebea5d9a81b2d8743b2d27be90b74685cd42a7b6