Rollback ref update when global-refdb update failure
Reference update contains two updates: local repository update
and global-refdb. Failure of the global-refdb update cause split
brain because local repository is ahead of the global-refdb. To
mitigate the problem, rollback local repository update when
global-refdb update fails.

Also for split brain we always rollback the ref-update and return
LOCK_FAILURE even if the enforcement policy is not set to REQUIRED.

Bug: Issue 14028
Change-Id: Ia7cc54a5b32fa2997cfb4673cb768fe43107b02a