Use a simpler named binding for the LocalDiskRepositoryManager
This partly reverts commit 0db559343d648180cb25d340c72f1e479165fb0c
and introduces a simpler named binding using the simple class name
of the LocalDiskRepositoryManager.

Allows to have a zero-configuration chaining between different
repository managers, using the optional binding technique.

Also, does not require anymore changes in other dependent plugins
such as multi-site and high-availability, because the absence of a
named binding is already managed with a fallback to the previous
legacy binding to the LocalDiskRepositoryManager class.

Bump the global-refdb version to 3.4.4.4.

Change-Id: I95ca0c2f9281384d29494c4594699a1c1add29ca