Make SharedRefDatabaseWrapper a singleton
The SharedRefDatabaseWrapper has optional injections on one of its
fields (sharedRefDbDynamicItem) which is resolved at the multi-site
and high-availability plugins startup phase.

The initial instance created during the plugin startup was injected
correctly thanks to the explicit call to inject fields; however, because
the SharedRefDatabaseWrapper was not a singleton, the subsequent
injections were done without the optional field, causing unexpected
behaviours.

Having a common singleton everywhere assures that the members injection
performed at startup would apply to all uses of SharedRefDatabaseWrapper.

Change-Id: Idb32bbd179629221d8d21c2986a1ad4b9f9077bf