Make DynamicItem<GlobalRefDatabase> a mandatory injection
The DynamicItem<GlobalRefDatabase> was defined as an optional injection
to workaround some issues in the Gerrit libModule setup between the
sysInjector and dbInjector when merging stable-2.16 to stable-3.0 in
I5be1200f97.

The workaround is no longer needed in stable-3.4 and therefore it
makes sense to remove it. That also allows to simplify the way the
SharedRefDatabaseWrapper is used in high-availability and multi-site
plugins.

Change-Id: If2bf3f2ec0c2b58dbb152c2d1af974bf6becaaf2